refactor: Security 관련 class 정리
