+++ perfume-api/src/main/java/io/perfume/api/common/jwt/JwtConfiguration.java
@@ -1,4 +1,4 @@
-package io.perfume.api.common.config;
+package io.perfume.api.common.jwt;
 
 import encryptor.impl.JwtUtil;
 import io.perfume.api.common.jwt.JwtProperties;
