+++ perfume-api/src/main/java/io/perfume/api/common/signIn/CustomUserDetailsService.java
@@ -1,4 +1,4 @@
-package io.perfume.api.common.jwt;
+package io.perfume.api.common.signIn;
 
 import io.perfume.api.user.application.port.out.UserQueryRepository;
 import org.springframework.security.core.userdetails.UserDetails;
